+++ datadog_checks_dev/datadog_checks/dev/tooling/utils.py
@@ -613,6 +613,19 @@ def has_logs(check):
     return False
 
 
+def is_metric_in_metadata_file(metric, check):
+    """
+    Return True if `metric` is listed in the check's `metadata.csv` file, False otherwise.
+    """
+    metadata_file = get_metadata_file(check)
+    if not os.path.isfile(metadata_file):
+        return False
+    for _, row in read_metadata_rows(metadata_file):
+        if row['metric_name'] == metric:
+            return True
+    return False
+
+
 def find_legacy_signature(check):
     """
     Validate that the given check does not use the legacy agent signature (contains agentConfig)
